updates to EmployeeService and tests
changed addSchedule and removeSchedule to private helper methods for addSchedules() and removeSchedules(), which have identical functionality.
Added a check to make sure we're not adding an employeeSchedule to employee that the employee already has (i.e. same date and shift, or same id) addSchedules fails otherwise
added tests to maintain 100% code coverage