removed unnecessary tests
this commit I'm leaving in a single commit because it's a drastic change.

So whether or not these tests are in the class or not, the test coverage reported by eclemma remains the same. That's because these tests test something that's already covered by another test.

A lot of these are just error checks on things like null/empty/spaces names when the method already calls getClass.

I also removed all empty name tests, since the spaces test does the exact same thing. If either of them fail, we know that the "name.trim().length() == 0" part of the code failed. Neither one tells us anything that the other doesn't already.

Unit testing -> It is given that any method called inside the method you're testing will work perfectly. Unit testing doesn't care if the getClass or delete method does something correctly or not. If it doesn't, then the error is in that method, not in the one we're testing.

When you call getOwner for example, you will always get a valid owner class. We know this because getOwner throws an error whenever it cannot get a valid owner class, and because of the "given", we assert that getOwner will work perfectly. Therefore, it will get an owner class.

So you don't need to test for cases where it doesn't.