added image to ItemServices
unfortunately, there's no way to verify if an image is valid or not in JPA. Well...there probably is, but it's much easier to do in javascript. Especially since the frontend will be responsible for communicating with external subsystems (e.g. delivery service, payment systems)

So, for error checks I simply check that the input is not null. this way, when handling it in javascript we are guaranteed a string of some kind when we call item.getImage().

Helps avoid the unpredictability javascript might have.